2nd Draft (Update) of Add Time Token Option
