--- src/app/token-options/spend-for-additional-assignment-time-request.ts
@@ -1,4 +1,0 @@
-import { TokenATMRequest } from './token-atm-request';
-import type { SpendForAdditionalAssignmentTimeTokenOption } from 'app/token-options/spend-for-additional-assignment-time-token-option';
-
-export class SpendForAdditionalAssignmentTimeRequest extends TokenATMRequest<SpendForAdditionalAssignmentTimeTokenOption> {}
